Changed ncap and added boundary conversion for intervals
Apparently ncap=60000 is the highest value possible before the algorithm breaks. So this is now the hardcoded value because it gives
the greatest accuracy. It can be decreased for speedup but then accuracy of the coefficients goes down.
Also n cannot be higher than ncap, a check for that should be implemented at a later stage.
Furthermore when h^2 is generated from J(omega), the interval boundaries on which h^2 is defined must be the boundaries of the
interval on which J is defined devided by g. This was also added to the code.