Rewrote tedopa/tedopa.py
Realized that the code wouldn't work the way it was written before, so
changed the way the algorithm works (tmps needs matrices as input and
does not work with mpas). Next step will be to adjust tedopa/tmps.py to
allow for two-site operators to start at the second site of the chain.
Then the code in tedopa/tedopa.py can be finished.